Fix memory leak when there is only one run.
Since the memory was released at the end of Merge, it ended up never being
released when Merge exited early (which is to say when there was less than
2 runs). I moved the memory release from the end of Merge to the end of
Sort to make sure that it was always released, and also made sure that the
memory was set to nullptr when released to avoid potential problem if the
algorithm tries to release the memory twice (I also took the liberty to
change every NULL to nullptr).

To be honest, this fix isn't enough to handle memory leaks when exceptions
are thrown but at least it makes the code safer when no exception is
thrown. You should probably use std::unique_ptr or another RAII guard to
make sure that there are no memory leaks, even when an exception is
thrown.